requirements.txt
# --- Core App Requirements ---
streamlit
openai>=1.0.0
requests>=2.27
python-dotenv

# --- Firebase Auth ---
pyrebase4
firebase-admin

# --- Voice & Audio ---
PyJWT
pydub
SpeechRecognition

# --- Plotting (Clarity radar chart) ---
matplotlib

# --- Google Cloud (Firestore) ---
google-cloud-firestore
google-auth

# --- Misc Helpers ---
setuptools
